requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
requests==2.28.2
selenium==4.7.2
validators==0.20.0
pytest==7.2.1

